fix(charts): timestamps and formatted values in tooltips in charts
